Fix CircularBuffer inner constructor

+++ src/circular_buffer.jl
@@ -14,14 +14,14 @@ mutable struct CircularBuffer{T} <: AbstractVector{T}
     length::Int
     buffer::Vector{T}
 
-    function CircularBuffer{T}(f,len,buf) where {T}
+    function CircularBuffer{T}(f::Int, len::Int, buf) where {T}
         f <= length(buf) || throw(ArgumentError("Value of 'first' must be inbounds of buffer"))
         len <= length(buf) || throw(ArgumentError("Value of 'length' must be <= length of buffer"))
         return new{T}(length(buf), f, len, buf)
     end
 
     # Convert any `Integer` to whatever `Int` is on the relevant machine
-    CircularBuffer{T}(f::Integer, len::Integer, buf::Integer) where {T} = CircularBuffer{T}(Int(f), Int(len), Int(buf))
+    CircularBuffer{T}(f::Integer, len::Integer, buf) where {T} = CircularBuffer{T}(Int(f), Int(len), buf)
 end
 
 function CircularBuffer{T}(iter, capacity::Integer) where {T}
